moved seed file generator to db folder and updated references, also deleted demo.js  as no longer needed
