+++ db/seeders/20220411165904-review-data.js
@@ -1,14 +1,14 @@
-'use strict';
+"use strict";
 
 // import array of object records
-const { reviewSeed } = require('../../public/javascripts/tables-seed-data');
+const { reviewSeed } = require("../tables-seed-data");
 
 module.exports = {
-  up: (queryInterface, Sequelize) => {
-    return queryInterface.bulkInsert('Reviews', reviewSeed, {});
-  },
+    up: (queryInterface, Sequelize) => {
+        return queryInterface.bulkInsert("Reviews", reviewSeed, {});
+    },
 
-  down: (queryInterface, Sequelize) => {
-    return queryInterface.bulkDelete('Reviews', null, {});
-  }
+    down: (queryInterface, Sequelize) => {
+        return queryInterface.bulkDelete("Reviews", null, {});
+    },
 };
